Subtitle: Skip subtitle only remain 33ms display time [1/1]
PD#SWPL-192698

Problem:
When zapping subtitles, the subtitle is posted to display if
which is only remain 33ms display time, it is a bad experience
that the subtitle flashes away on the screen.

Solution:
Stop display such subtitle if the subtitle's display time is
less than 33ms.

Verify:
Verified On Android U

Change-Id: I58d21ba350995fb4b0b1225c311ff8b5e576d3ed
Signed-off-by: hongliang.zheng <hongliang.zheng@amlogic.com>